Tikhonov+general refactor to better handle params.
Previous code only allowed one Tikhonov regularization parameter for
rotation and translation, but this does not work well because the
diagonal Hessian elements take on different orders of magnitude in
different levels. So I had to accomodate having separate params for each
level (which also doesn't work so far, but whatever). Doing so, I
chose to make a larger refactor that wraps all these params in a struct.